Added ffmpeg check to install.sh

+++ install.sh
@@ -8,7 +8,8 @@ EXTENSION_API_PATH=../gdextension-api-files/gd4.4_linux_extension_api.json
 echo "PREREQUISITES"
 echo "- CPP-COMPILER: g++"
 echo "- BUILD TOOLS: scons, cmake"
-echo "- OTHER: winget, git"
+echo "- OTHER: wget, git"
+echo "- LIBRARIES: ffmpeg, libavcodec"
 
 
 echo "STEP 1 : BUILD OPENCV"
@@ -18,6 +19,10 @@ unzip opencv.zip && unzip opencv_contrib.zip && rm -vf opencv.zip && rm -vf open
 mkdir {build,opencv}
 cd build
 cmake -DCMAKE_INSTALL_PREFIX=../opencv -DOPENCV_EXTRA_MODULES_PATH=../opencv_contrib-$OPENCV_VERSION/modules -DBUILD_LIST=$OPENCV_BUILD_LIST -DWITH_FFMPEG=ON ../opencv-$OPENCV_VERSION
+if [ $(cmake -LAH | grep -c ".*FFMPEG:.*YES") -nw 1 ]; then
+	echo "WARNING: FAILED TO PROPERLY BUILD WITH FFMPEG"
+	exit
+fi
 cmake --build . --target install
 cd ..
 rm -vrf opencv_contrib-$OPENCV_VERSION && rm -vrf opencv-$OPENCV_VERSION
